Changes for restricting delete student or making student inactive based on configuration.

+++ dep-studentmgmt-ctrl/src/com/ctb/control/studentManagement/StudentManagementImpl.java
@@ -1539,6 +1539,10 @@ public class StudentManagementImpl implements StudentManagement
 			}
 
 			com.ctb.bean.testAdmin.OrgNodeStudent [] orgNodeStus = orgNodeStudents.getOrgNodeStudentForStudentAtAndBelowOrgNodes(studentId, SQLutils.generateSQLCriteria(findInColumn,topOrgNodeIds));
+			
+			// Check whether the customer has configuration for restricting delete student or not.
+			int canDeleteStudent = studentManagement.isConfigurationPresent(studentId,"Disable_Delete_Move_Student");
+			
 			for (int i=0; orgNodeStus!=null && i< orgNodeStus.length; i++) {
 				com.ctb.bean.testAdmin.OrgNodeStudent oldOrgNodeInDB = orgNodeStus[i];
 				boolean foundInNewOrgNodes = newOrgNodeHash.containsKey(oldOrgNodeInDB.getOrgNodeId());
@@ -1548,12 +1552,16 @@ public class StudentManagementImpl implements StudentManagement
 					newOrgNodeHash.remove(oldOrgNodeInDB.getOrgNodeId());                              
 				}
 				else { //delete or deactivate
-					Integer rosterCount = testRosters.getRosterCountForStudentAndOrgNode(studentId, oldOrgNodeInDB.getOrgNodeId());
-					if (rosterCount.intValue() >0) {
-						orgNodeStudents.deactivateOrgNodeStudentForStudentAndOrgNode(studentId, oldOrgNodeInDB.getOrgNodeId());
-					}
-					else {
-						orgNodeStudents.deleteOrgNodeStudentForStudentAndOrgNode(studentId, oldOrgNodeInDB.getOrgNodeId());
+					if(canDeleteStudent < 1) { // If configuration Disable_Delete_Move_Student is present then student can neither be deleted nor deactivated
+						Integer rosterCount = testRosters.getRosterCountForStudentAndOrgNode(studentId, oldOrgNodeInDB.getOrgNodeId());
+						if (rosterCount.intValue() >0) {
+							orgNodeStudents.deactivateOrgNodeStudentForStudentAndOrgNode(studentId, oldOrgNodeInDB.getOrgNodeId());
+						}
+						else {
+							orgNodeStudents.deleteOrgNodeStudentForStudentAndOrgNode(studentId, oldOrgNodeInDB.getOrgNodeId());
+						}
+					} else {
+						throw new CTBBusinessException("Failed to update student.");
 					}
 				}
 			}
@@ -1684,7 +1692,14 @@ public class StudentManagementImpl implements StudentManagement
 	{
 		validator.validateStudent(userName, studentId, "StudentManagementImpl.deleteStudent");
 		try {
-
+			// Check whether the customer has configuration for restricting delete student or not.
+			int canDeleteStudent = studentManagement.isConfigurationPresent(studentId,"Disable_Delete_Move_Student");
+			if(canDeleteStudent > 0) {
+				CTBBusinessException be = new CTBBusinessException( 
+						"You cannot delete this student.");
+				throw be;
+			}
+			
 			RosterElement [] rosters = studentManagement.getRosterElementsForStudent(studentId);
 
 			int deleteStatus = 0;	// 0: do not touch student (default)
